refactor: int -> long

+++ src/main/java/com/scoreboard/mapper/MatchesPageMapper.java
@@ -7,14 +7,14 @@ import java.util.List;
 
 public class MatchesPageMapper {
 
-    public MatchesPage map(int pageNumber, List<Match> matches,
-                           int totalPages, String filterByPlayerName) {
+    public MatchesPage map(long pageNumber, List<Match> matches,
+                           long totalPages, String filterByPlayerName) {
         boolean hasFilter = filterByPlayerName != null && !filterByPlayerName.trim().isEmpty();
         boolean shouldShowPagination = totalPages > 1;
         boolean hasPrevious = pageNumber > 1;
         boolean hasNext = pageNumber < totalPages;
-        int previousPage = hasPrevious ? pageNumber - 1 : 1;
-        int nextPage = hasNext ? pageNumber + 1 : Math.max(1, totalPages);
+        long previousPage = hasPrevious ? pageNumber - 1 : 1;
+        long nextPage = hasNext ? pageNumber + 1 : Math.max(1, totalPages);
 
         return MatchesPage.builder()
                 .pageNumber(pageNumber)
